+++ StarLevelSystem/StarLevelSystem.cs
@@ -22,7 +22,7 @@ namespace StarLevelSystem
     {
         public const string PluginGUID = "MidnightsFX.StarLevelSystem";
         public const string PluginName = "StarLevelSystem";
-        public const string PluginVersion = "0.17.11";
+        public const string PluginVersion = "0.17.12";
 
         public ValConfig cfg;
         // Use this class to add your own localization to the game